Add mockCustomizationContext for tests
