+++ src/plugins/discover/public/__mocks__/discover_state.mock.ts
@@ -10,6 +10,7 @@ import { getDiscoverStateContainer } from '../application/main/services/discover
 import { savedSearchMockWithTimeField, savedSearchMock } from './saved_search';
 import { discoverServiceMock } from './services';
 import { SavedSearch } from '@kbn/saved-search-plugin/public';
+import { mockCustomizationContext } from '../customizations/__mocks__/customization_context';
 
 export function getDiscoverStateMock({
   isTimeBased = true,
@@ -23,13 +24,7 @@ export function getDiscoverStateMock({
   const container = getDiscoverStateContainer({
     services: discoverServiceMock,
     history,
-    customizationContext: {
-      displayMode: 'standalone',
-      inlineTopNav: {
-        enabled: false,
-        showLogsExplorerTabs: false,
-      },
-    },
+    customizationContext: mockCustomizationContext,
   });
   container.savedSearchState.set(
     savedSearch ? savedSearch : isTimeBased ? savedSearchMockWithTimeField : savedSearchMock
