+++ src/plugins/discover/public/application/discover_router.test.tsx
@@ -23,7 +23,7 @@ import { ContextAppRoute } from './context';
 import { createProfileRegistry } from '../customizations/profile_registry';
 import { addProfile } from '../../common/customizations';
 import { NotFoundRoute } from './not_found';
-import type { DiscoverCustomizationContext } from './types';
+import { mockCustomizationContext } from '../customizations/__mocks__/customization_context';
 
 let mockProfile: string | undefined;
 
@@ -49,18 +49,10 @@ const gatherRoutes = (wrapper: ShallowWrapper) => {
   });
 };
 
-const customizationContext: DiscoverCustomizationContext = {
-  displayMode: 'standalone',
-  inlineTopNav: {
-    enabled: false,
-    showLogsExplorerTabs: false,
-  },
-};
-
 const props: DiscoverRoutesProps = {
   isDev: false,
   customizationCallbacks: [],
-  customizationContext,
+  customizationContext: mockCustomizationContext,
 };
 
 describe('DiscoverRoutes', () => {
@@ -164,7 +156,7 @@ describe('CustomDiscoverRoutes', () => {
     const component = shallow(
       <CustomDiscoverRoutes
         profileRegistry={profileRegistry}
-        customizationContext={customizationContext}
+        customizationContext={mockCustomizationContext}
         isDev={props.isDev}
       />
     );
@@ -172,7 +164,7 @@ describe('CustomDiscoverRoutes', () => {
       <DiscoverRoutes
         prefix={addProfile('', mockProfile)}
         customizationCallbacks={callbacks}
-        customizationContext={customizationContext}
+        customizationContext={mockCustomizationContext}
         isDev={props.isDev}
       />
     );
@@ -183,7 +175,7 @@ describe('CustomDiscoverRoutes', () => {
     const component = shallow(
       <CustomDiscoverRoutes
         profileRegistry={profileRegistry}
-        customizationContext={customizationContext}
+        customizationContext={mockCustomizationContext}
         isDev={props.isDev}
       />
     );
@@ -202,7 +194,7 @@ describe('DiscoverRouter', () => {
         services={mockDiscoverServices}
         history={history}
         profileRegistry={profileRegistry}
-        customizationContext={customizationContext}
+        customizationContext={mockCustomizationContext}
         isDev={props.isDev}
       />
     );
@@ -213,7 +205,7 @@ describe('DiscoverRouter', () => {
     expect(pathMap['/']).toMatchObject(
       <DiscoverRoutes
         customizationCallbacks={callbacks}
-        customizationContext={customizationContext}
+        customizationContext={mockCustomizationContext}
         isDev={props.isDev}
       />
     );
@@ -223,7 +215,7 @@ describe('DiscoverRouter', () => {
     expect(pathMap[profilePath]).toMatchObject(
       <CustomDiscoverRoutes
         profileRegistry={profileRegistry}
-        customizationContext={customizationContext}
+        customizationContext={mockCustomizationContext}
         isDev={props.isDev}
       />
     );
