+++ src/plugins/discover/public/application/main/components/top_nav/on_save_search.test.tsx
@@ -18,6 +18,7 @@ import { ReactElement } from 'react';
 import { discoverServiceMock } from '../../../../__mocks__/services';
 import { SavedSearch } from '@kbn/saved-search-plugin/public';
 import { createBrowserHistory } from 'history';
+import { mockCustomizationContext } from '../../../../customizations/__mocks__/customization_context';
 
 function getStateContainer({ dataView }: { dataView?: DataView } = {}) {
   const savedSearch = savedSearchMock;
@@ -25,13 +26,7 @@ function getStateContainer({ dataView }: { dataView?: DataView } = {}) {
   const stateContainer = getDiscoverStateContainer({
     services: discoverServiceMock,
     history,
-    customizationContext: {
-      displayMode: 'standalone',
-      inlineTopNav: {
-        enabled: false,
-        showLogsExplorerTabs: false,
-      },
-    },
+    customizationContext: mockCustomizationContext,
   });
   stateContainer.savedSearchState.set(savedSearch);
   stateContainer.appState.getState = jest.fn(() => ({
