+++ src/plugins/discover/public/application/main/discover_main_route.tsx
@@ -35,10 +35,10 @@ import { useAlertResultsToast } from './hooks/use_alert_results_toast';
 import { DiscoverMainProvider } from './services/discover_state_provider';
 import {
   CustomizationCallback,
+  DiscoverCustomizationContext,
   DiscoverCustomizationProvider,
   useDiscoverCustomizationService,
 } from '../../customizations';
-import type { DiscoverCustomizationContext } from '../types';
 import { DiscoverTopNavInline } from './components/top_nav/discover_topnav_inline';
 import { isTextBasedQuery } from './utils/is_text_based_query';
 import { DiscoverStateContainer, LoadParams } from './services/discover_state';
